Split routes.tsx into multiple files.
